Refactor licenses to many files
